Add files via upload
This project contains the Ola-bike dataset, cleaning and visualisation using pandas and plotly. encoding and scaling of the dataset. application of different Regression models for machine learning. 
Deployment of the encoded data and the best Regression models using Streamlit (VS code) for my prediction app.
Creating a dashboard for visualisation and interaction with the cleaned data.